Created CompanyValidationRequest, CompanyValidationResponse, VacancyValidationRequest and VacancyValidationResponse. Created application.yaml

+++ src/main/java/com/tvo/technologies/saferecruitment/model/user/User.java
@@ -8,7 +8,7 @@ public class User {
     private String password;
     private String name;
     private String surname;
-    private float experience;
+    private int experience;
     private EducationLevel education;
     private String targetPosition;
     private String additionalInformation;
